Sounds cropped
We are experiencing problems with the sounds. When a key is pressed,
the sound has to play till the end (around 4 seconds) until it can be
pressed again. This has proved to be very hard to work around. We tried
looking into something called “sound pool”, but this method is very
poorly described for c#, so we right now we are sticking with the sound
player method, and have now cut the last 3 seconds of each sound off.