UML added
A UML constructed in Violet. Hopefully accurate enough to give an
impression of the “blueprint” of the project. This is the third version
of our UML, and has a much more specific in the description, and
produced closer to how the actual code is performed. The initial
version was more of a “pseudo” UML, and the second more of a mix
between pseudo and the actual program.